Refactoring code and fixing error
