--- delivery/src/components/DescriptionCover.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-const DescriptionCover = styled.img.attrs({
-    src: "img/picpizzaaa.jpg"
-})`
-    width: 100%;
-`;
-
-export default DescriptionCover;
